+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add Control Plane drainer controller.
+
 ### Fixed
 
 - Fix aws operator policy for latest node pools version.